+++ framework/db-http/src/main/java/is/codion/framework/db/http/JsonHttpEntityConnection.java
@@ -316,7 +316,7 @@ final class JsonHttpEntityConnection extends AbstractHttpEntityConnection {
 	}
 
 	@Override
-	public <T, P, R> R report(ReportType<T, P, R> reportType, P parameter) {
+	public <P, R> R report(ReportType<P, R> reportType, P parameter) {
 		requireNonNull(reportType);
 		synchronized (transport) {
 			try {